feat(config): Add separate /reanimated entry point for testing reanimated configs easier
