+++ apps/kitchen-sink/package.json
@@ -42,6 +42,7 @@
     "react-dom": "^18.2.0",
     "react-native": "^0.71.4",
     "react-native-gesture-handler": "^2.8.0",
+    "react-native-reanimated": "^3.0.2",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
     "react-native-svg": "13.4.0",
