+++ apps/kitchen-sink/src/tamagui.config.ts
@@ -1,7 +1,9 @@
-import { config } from '@tamagui/config'
+import { config as configType } from '@tamagui/config'
+import { config } from '@tamagui/config/reanimated'
 import { createTamagui } from 'tamagui'
 
-const tamaConf = createTamagui(config)
+// doing tricks to get types right
+const tamaConf = createTamagui(config as typeof configType)
 
 export type Conf = typeof tamaConf
