+++ apps/sandbox/package.json
@@ -25,7 +25,7 @@
     "@tamagui/vite-plugin-reanimated": "1.9.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-native-reanimated": "^2.14.4",
+    "react-native-reanimated": "^3.0.2",
     "react-native-web": "^0.18.3",
     "solito": "^2.3.0",
     "tamagui": "1.9.1"
