+++ packages/animations-reanimated/package.json
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@tamagui/build": "1.9.1",
-    "react-native-reanimated": "^2.14.4"
+    "react-native-reanimated": "^3.0.2"
   },
   "scripts": {
     "build": "tamagui-build",
