+++ packages/config/reanimated.js
@@ -1,0 +1,2 @@
+// for react native... and typescript... because type module breaks things
+module.exports = require('./dist/cjs/reanimated')
