+++ packages/config/src/reanimated.ts
@@ -1,8 +1,7 @@
 export * from './media'
 export * from './createGenericFont'
-export { animations } from './animations'
 
-import { animations } from './animations'
+import { animations } from './animations.reanimated'
 import { createTamaguiConfig } from './createTamaguiConfig'
 
 export const config = createTamaguiConfig(animations)
